requirements.txt
# Зависимости для orders_backend (Django + DRF)
django~=5.0
djangorestframework~=3.14.0
django-filter>=24.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
PyYAML~=6.0.0


# requests~=2.31.0
# ujson>=5.10.0
# django-rest-passwordreset>=1.3.0
